+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to `Credo` will be documented in this file.
 
 ## [0.6.0](https://github.com/rougin/credo/compare/v0.5.0...v0.6.0) - Unreleased
 
+### Added
+- `$pagee` for customizing pagination in `PaginateTrait`
+- Methods in `Repository` related to CRUD operations
+
 ### Changed
 - `CredoTrait::credo` returns the `Credo` instance
 - Code coverage from `Scrutinizer CI` to `Codecov`