+++ CHANGELOG.md
@@ -22,6 +22,9 @@ All notable changes to `Credo` will be documented in this file.
 ### Fixed
 - Getting offset value in `PaginateTrait`
 
+### Removed
+- `__call` magic method in `Repository`
+
 ## [0.5.0](https://github.com/rougin/credo/compare/v0.4.0...v0.5.0) - 2018-12-08
 
 > [!WARNING]